#1227 - Integrate and clean up PDFExtract code
- Move content of README.md into package-info.java
- Add license header to files
- Update NOTICE and LICENSE file
