+++ inception-pdf-editor/pom.xml
@@ -261,7 +261,6 @@
                   <!-- These upstream files do not contain license headers -->
                   <exclude>src/main/js/pdfanno/**/*</exclude>
                   <exclude>cache/**/*</exclude>
-                  <exclude>src/main/java/paperai/</exclude>
                 </excludes>
               </configuration>
             </execution>
