+++ inception-pdf-editor/src/main/java/de/tudarmstadt/ukp/inception/pdfeditor/pdfextract/ImageExtractor.java
@@ -1,4 +1,27 @@
-package paperai.pdfextract;
+/*
+ * The MIT License (MIT)
+ * 
+ * Copyright (c) 2017 Hiroyuki Shindo
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package de.tudarmstadt.ukp.inception.pdfeditor.pdfextract;
 
 import java.awt.image.RenderedImage;
 import java.io.File;
@@ -34,8 +57,9 @@ public class ImageExtractor
     {
         File inFile = new File(args[0]);
         HashMap<String, String> map = new HashMap<>();
-        for (int i = 1; i < args.length; i += 2)
+        for (int i = 1; i < args.length; i += 2) {
             map.put(args[i], args[i + 1]);
+        }
         String o = map.containsKey("-o") ? map.get("-o") : "";
         String outDir = o.isEmpty() ? inFile.getParent() : (new File(o)).getAbsolutePath();
         int dpi = map.containsKey("-dpi") ? Integer.parseInt(map.get("-dpi")) : 300;
@@ -47,8 +71,9 @@ public class ImageExtractor
                 }
             }
         }
-        else
+        else {
             processFile(inFile, dpi, outDir);
+        }
     }
 
     static void processFile(File inFile, int dpi, String outDir) throws IOException
