--- inception-pdf-editor/src/main/java/paperai/pdfextract/DrawOperator.java
@@ -1,13 +1,0 @@
-package paperai.pdfextract;
-
-public class DrawOperator
-{
-    String type;
-    float[] values;
-
-    public DrawOperator(String type, float[] values)
-    {
-        this.type = type;
-        this.values = values;
-    }
-}
