--- inception-pdf-editor/src/main/java/paperai/pdfextract/ImageOperator.java
@@ -1,17 +1,0 @@
-package paperai.pdfextract;
-
-public class ImageOperator
-{
-    float x;
-    float y;
-    float w;
-    float h;
-
-    public ImageOperator(float x, float y, float w, float h)
-    {
-        this.x = x;
-        this.y = y;
-        this.w = w;
-        this.h = h;
-    }
-}
